GH Update: adding mapping os sponsors
